+++ rend3-routine/shaders/src/lighting/pixel.glsl
@@ -235,7 +235,7 @@ PixelData get_per_pixel_data_sampled(MATERIAL_TYPE material, sampler s) {
     return pixel;
 }
 
-#ifdef GPU_MODE
+#ifdef GPU_DRIVEN
 PixelData get_per_pixel_data(MATERIAL_TYPE material) {
     if (MATERIAL_FLAG(FLAGS_NEAREST)) {
         return get_per_pixel_data_sampled(material, nearest_sampler);
@@ -245,8 +245,8 @@ PixelData get_per_pixel_data(MATERIAL_TYPE material) {
 }
 #endif
 
-#ifdef CPU_MODE
-// In CPU mode the primary sampler gets switched out for what we need, so we don't switch in the shader.
+#ifdef CPU_DRIVEN
+// In the CpuDriven profile the primary sampler gets switched out for what we need, so we don't switch in the shader.
 // This is because OpenGL can't deal with any texture being used with multiple different samplers. 
 PixelData get_per_pixel_data(MATERIAL_TYPE material) {
     return get_per_pixel_data_sampled(material, primary_sampler);