+++ rend3-routine/src/common/shaders.rs
@@ -1,38 +1,38 @@
 use std::borrow::Cow;
 
-use rend3::RendererMode;
+use rend3::RendererProfile;
 use wgpu::{Device, ShaderModule, ShaderModuleDescriptor, ShaderModuleDescriptorSpirV, ShaderSource};
 
 use crate::shaders::{SPIRV_SHADERS, WGSL_SHADERS};
 
-/// In cpu-mode, creates a checked wgsl shader, in gpu-mode creates a
+/// When CpuDriven, creates a checked wgsl shader, when CpuDriven creates a
 /// passthrough SPIRV shader.
 ///
 /// # Safety
 ///
 /// The shader must be valid, match all the respective definitions, and
 /// otherwise meet wgpu's validation requirements
-pub unsafe fn mode_safe_shader(
+pub unsafe fn profile_safe_shader(
     device: &Device,
-    mode: RendererMode,
+    profile: RendererProfile,
     label: &str,
     cpu_source: &str,
     gpu_source: &str,
 ) -> ShaderModule {
-    let shader_dir = match mode {
-        RendererMode::CpuPowered => &WGSL_SHADERS,
-        RendererMode::GpuPowered => &SPIRV_SHADERS,
+    let shader_dir = match profile {
+        RendererProfile::CpuDriven => &WGSL_SHADERS,
+        RendererProfile::GpuDriven => &SPIRV_SHADERS,
     };
 
     let source = shader_dir
-        .get_file(match mode {
-            RendererMode::CpuPowered => cpu_source,
-            RendererMode::GpuPowered => gpu_source,
+        .get_file(match profile {
+            RendererProfile::CpuDriven => cpu_source,
+            RendererProfile::GpuDriven => gpu_source,
         })
         .unwrap()
         .contents();
 
-    let use_unsafe = mode == RendererMode::GpuPowered;
+    let use_unsafe = profile == RendererProfile::GpuDriven;
 
     match use_unsafe {
         false => device.create_shader_module(&ShaderModuleDescriptor {