+++ rend3/src/lib.rs
@@ -1,27 +1,102 @@
 //! Easy to use, customizable, efficient 3D renderer library built on wgpu.
 //!
-//! Library is under active development. While internals will likely change
-//! quite a bit, the external api will only experience minor changes as features
-//! are added.
+//! Library is under active development. While internals are might change in the
+//! future, the external api remains stable, with only minor changes occuring as
+//! features are added.
 //!
-//! To use rend3 add the following to your Cargo.toml:
+//! # Examples
 //!
-//! ```text
-//! rend3 = "0.2.0"
-//! ```
+//! Take a look at the [examples] for getting started with the api. The examples
+//! will show how the core library and helper crates can be used.
 //!
-//! # Screenshots
+//! [examples]: https://github.com/BVE-Reborn/rend3/tree/trunk/examples
+//!
+//! ### Screenshots
+//!
+//! These screenshots are from the scene-viewer example.
 //!
 //! ![scifi-base](https://raw.githubusercontent.com/BVE-Reborn/rend3/trunk/examples/scene-viewer/scifi-base.jpg)
 //! ![example](https://raw.githubusercontent.com/BVE-Reborn/rend3/trunk/examples/scene-viewer/screenshot.jpg)
 //! ![bistro](https://raw.githubusercontent.com/BVE-Reborn/rend3/trunk/examples/scene-viewer/bistro.jpg)
 //! ![emerald-square](https://raw.githubusercontent.com/BVE-Reborn/rend3/trunk/examples/scene-viewer/emerald-square.jpg)
 //!
-//! # Examples
+//! # Crates
 //!
-//! Take a look at the [examples] getting started with the api.
+//! The `rend3` ecosystem is composed of a couple core crates which provide most
+//! of the functionality and exensibility to the library, extension crates, and
+//! integration crates
 //!
-//! [examples]: https://github.com/BVE-Reborn/rend3/tree/trunk/examples
+//! ### Core
+//!
+//! - `rend3`: The core crate. Performs all handling of world data, provides the
+//!   Renderer and RenderGraph and defines vocabulary types.
+//! - `rend3-routine`: Implementation of various "Render Routines" on top of the
+//!   RenderGraph. Also provides for re-usable graphics work. Provides PBR
+//!   rendering, Skyboxes, Shadow Rendering, and Tonemapping.
+//!
+//! ### Extensions
+//!
+//! There are extension crates that are not required, but provide pre-made bits
+//! of useful code that I would recommend using.
+//!
+//! - `rend3-framework`: Vastly simplifies correct handling of the window and
+//!   surface across platforms.
+//! - `rend3-gltf`: Modular gltf file and scene loader.
+//!
+//! ### Integration
+//!
+//! Integration with other external libraries are also offered. Due to external
+//! dependencies, the versions of these may increase at a much higher rate than
+//! the rest of the ecosystem.
+//!
+//! - `rend3-egui`: Integration with the [egui](https://github.com/emilk/egui)
+//!   immediate mode gui.
+//! - `rend3-imgui`: Integration with the [imgui](https://github.com/ocornut/imgui)
+//!   immediate mode gui.
+//!
+//! # Features and Platform Support
+//!
+//! rend3 supports two different rendering profiles one for speed and one for
+//! compatibility.
+//!
+//! ### Profile Features
+//!
+//! The modern profile not only offloads a lot more work to the gpu, it can do more
+//! aggressive performance optimizations including only drawing exactly the triangles that are needed
+//!
+//! | Profile   | Texture Access | Object Culling | Triangle Culling | Draw Calls          |
+//! |:----------|----------------|----------------|------------------|---------------------|
+//! | GpuDriven | Bindless       | On GPU         | On GPU           | Merged Indirect     |
+//! | CpuDriven | Bound          | On CPU         | ❌                | Instanced Direct    |
+//!
+//! ### Profile Support
+//!
+//! The following table shows support of various profiles on various apis/platforms. This will
+//! hopefully help you judge what your target demographic supports.
+//!
+//! | OS                   | API    | GPU                                        | GpuDriven | CpuDriven |
+//! |----------------------|--------|--------------------------------------------|:---------:|:---------:|
+//! | Windows 7+           | Vulkan | AMD / NVIDIA                               | ✅    | —       |
+//! |                      | Vulkan | Intel 6XXX+                                | ❌      | ✅     |
+//! |                      | Dx11   | Intel 2XXX+                                | ❌      | 🚧     |
+//! | Windows 10+          | Dx12   | Intel 6XXX+ / AMD GCN 2+ / NVIDIA 6XX+     | 🚧    | ✅     |
+//! | MacOS 10.13+ iOS 11+ | Metal  | Intel / Apple A13+ / M1+                   | ✅    | —       |
+//! |                      |        | Apple A9+                                  | ❌      | ✅     |
+//! | Linux                | Vulkan | Intel 6XXX+ / AMD GCN 2+  / NVIDIA 6XX+    | ✅    | —       |
+//! |                      |        | Intel 4XXX+                                | ❌      | ✅     |
+//! | Android              | Vulkan | All                                        | ❌      | ✅     |
+//!
+//! Footnotes:
+//! - ✅ Supported
+//! - 🚧 In Progress
+//! - ❌ Unsupported
+//! - — Modern Profile Used
+//! - Intel 6XXX = Skylake
+//! - Intel 4XXX = Haswell
+//! - Intel 2XXX = Sandy Bridge
+//! - AMD GCN 2 = Rx 200+, RX 5000+
+//! - Apple A9 = iPhone 6S, iPad 5th Gen
+//! - Apple A13 = iPhone 11, iPad 9th Gen
 //!
 //! # Purpose
 //!
@@ -36,19 +111,6 @@
 //! an advanced game or simulation nor care how you structure     your program.
 //! I do have plans for a `rend3-util` (or similar) crate that is a very basic
 //! framework for the second use case listed above.
-//!
-//! # Helper Crates
-//!
-//! This is the primary crate which holds the main [`Renderer`] struct. We have
-//! some other crates:
-//! - `rend3-gltf`: contains code to load from a .gltf or .glb file.
-//! - `rend3-routine`: contains render routines for drawing PBR-style objects.
-//!
-//! # GPU Culling
-//!
-//! On Vulkan and DX12 "gpu mode" is enabled by default, which uses modern
-//! bindless resources and gpu-based culling. This reduces CPU load and allows
-//! significantly more powerful culling.
 //!
 //! # Future Plans
 //!
@@ -126,11 +188,11 @@ pub mod util {
 
 pub mod graph;
 mod instruction;
-mod mode;
+mod profile;
 mod setup;
 mod surface;
 
-pub use mode::*;
+pub use profile::*;
 pub use renderer::{error::*, Renderer, RendererDataCore};
 pub use setup::*;
 pub use surface::*;